requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

numpy==1.22.4
numba
scipy
Unidecode
openjtalk>=0.3.0.dev2
jamo
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.12.0+cpu
